fix: Prevent immediate heal after standing up
The application would sometimes send a `/heal` command immediately after the player stood up from a non-standing state (e.g., healing). This was because the idle timer was not being reset upon this state change.

This commit introduces a `_previousPlayerStatus` field to track the player's status on each tick. When a transition from a non-standing state to a standing state is detected, the idle heal timer is reset. This prevents an unwanted `/heal` command from being sent.

This also fixes a build error caused by an implicit conversion from `uint` to `byte` when reading the player's status.

+++ Form1.cs
@@ -6217,7 +6217,7 @@ private string GetBestSpellTier(string buffType, string targetName)
                     }
                     _ELITEAPIPL.ThirdParty.SendString("/heal");
                 }
-                byte currentPlayerStatus = _ELITEAPIPL.Player.Status;
+                byte currentPlayerStatus = (byte)_ELITEAPIPL.Player.Status;
 
                 // If player just stood up from a non-standing state (like healing), reset the idle timer.
                 if (_previousPlayerStatus != (byte)Status.Standing && currentPlayerStatus == (byte)Status.Standing)
